+++ autokeras/backend/__init__.py
@@ -11,3 +11,11 @@ class Backend:
     @classmethod
     def get_image_transformer(cls, x_train, augment=None):
         return cls.backend.ImageDataTransformer(x_train, augment=augment)
+
+    @classmethod
+    def produce_model(cls, graph):
+        return cls.backend.produce_model(graph)
+
+    @classmethod
+    def get_model_trainer(cls, **kwargs):
+        return cls.backend.ModelTrainer(**kwargs)
