+++ autokeras/backend/torch/__init__.py
@@ -1,0 +1,3 @@
+from autokeras.backend.torch.model import produce_model
+from autokeras.backend.torch.data_transformer import ImageDataTransformer
+from autokeras.backend.torch.model_trainer import ModelTrainer
