+++ autokeras/image/gan.py
@@ -8,9 +8,9 @@ from copy import deepcopy
 from torchvision import utils as vutils
 from tqdm import tqdm
 
+from autokeras.backend import Backend
 from autokeras.constant import Constant
 from autokeras.nn.loss_function import binary_classification_loss
-from autokeras.backend.torch import ImageDataTransformer
 from autokeras.nn.model_trainer import ModelTrainerBase
 from autokeras.unsupervised import Unsupervised
 from autokeras.utils import get_device
@@ -56,7 +56,7 @@ class DCGAN(Unsupervised):
         """
         # input size stay the same, enable  cudnn optimization
         cudnn.benchmark = True
-        self.data_transformer = ImageDataTransformer(x_train, augment=self.augment)
+        self.data_transformer = Backend.get_image_transformer(x_train, augment=self.augment)
         train_dataloader = self.data_transformer.transform_train(x_train)
         GANModelTrainer(self.net_g,
                         self.net_d,
