+++ tests/test_search.py
@@ -16,7 +16,7 @@ from nas.random import RandomSearcher
 
 @patch('torch.multiprocessing.get_context', side_effect=MockProcess)
 @patch('autokeras.bayesian.transform', side_effect=simple_transform)
-@patch('autokeras.search.ModelTrainer.train_model', side_effect=mock_train)
+@patch('autokeras.backend.torch.model_trainer.ModelTrainer.train_model', side_effect=mock_train)
 def test_bayesian_searcher(_, _1, _2):
     train_data, test_data = get_classification_data_loaders()
     clean_dir(TEST_TEMP_DIR)
@@ -31,7 +31,7 @@ def test_bayesian_searcher(_, _1, _2):
 
 
 @patch('torch.multiprocessing.get_context', side_effect=MockProcess)
-@patch('autokeras.search.ModelTrainer.train_model', side_effect=mock_exception_handling_train)
+@patch('autokeras.backend.torch.model_trainer.ModelTrainer.train_model', side_effect=mock_exception_handling_train)
 def test_exception_handling(_, _2):
     train_data, test_data = get_classification_data_loaders()
     clean_dir(TEST_TEMP_DIR)
@@ -48,7 +48,7 @@ def test_exception_handling(_, _2):
 
 @patch('torch.multiprocessing.get_context', side_effect=MockProcess)
 @patch('autokeras.bayesian.transform', side_effect=simple_transform)
-@patch('autokeras.search.ModelTrainer.train_model', side_effect=mock_train)
+@patch('autokeras.backend.torch.model_trainer.ModelTrainer.train_model', side_effect=mock_train)
 @patch('autokeras.search.get_system', return_value=Constant.SYS_GOOGLE_COLAB)
 def test_bayesian_searcher_sp(_, _1, _2, _3):
     train_data, test_data = get_classification_data_loaders()
@@ -65,7 +65,7 @@ def test_bayesian_searcher_sp(_, _1, _2, _3):
 
 @patch('torch.multiprocessing.get_context', side_effect=MockProcess)
 @patch('autokeras.bayesian.transform', side_effect=simple_transform_mlp)
-@patch('autokeras.search.ModelTrainer.train_model', side_effect=mock_train)
+@patch('autokeras.backend.torch.model_trainer.ModelTrainer.train_model', side_effect=mock_train)
 def test_bayesian_searcher_mlp(_, _1, _2):
     train_data, test_data = get_classification_data_loaders_mlp()
     clean_dir(TEST_TEMP_DIR)
@@ -81,7 +81,7 @@ def test_bayesian_searcher_mlp(_, _1, _2):
 
 @patch('torch.multiprocessing.get_context', side_effect=MockProcess)
 @patch('autokeras.bayesian.transform', side_effect=simple_transform)
-@patch('autokeras.search.ModelTrainer.train_model', side_effect=mock_train)
+@patch('autokeras.backend.torch.model_trainer.ModelTrainer.train_model', side_effect=mock_train)
 def test_greedy_searcher(_, _1, _2):
     train_data, test_data = get_classification_data_loaders()
     clean_dir(TEST_TEMP_DIR)
@@ -95,7 +95,7 @@ def test_greedy_searcher(_, _1, _2):
 
 @patch('torch.multiprocessing.get_context', side_effect=MockProcess)
 @patch('autokeras.bayesian.transform', side_effect=simple_transform)
-@patch('autokeras.search.ModelTrainer.train_model', side_effect=mock_train)
+@patch('autokeras.backend.torch.model_trainer.ModelTrainer.train_model', side_effect=mock_train)
 @patch('autokeras.search.get_system', return_value=Constant.SYS_GOOGLE_COLAB)
 def test_greedy_searcher_sp(_, _1, _2, _3):
     train_data, test_data = get_classification_data_loaders()
@@ -110,7 +110,7 @@ def test_greedy_searcher_sp(_, _1, _2, _3):
 
 @patch('torch.multiprocessing.get_context', side_effect=MockProcess)
 @patch('autokeras.bayesian.transform', side_effect=simple_transform_mlp)
-@patch('autokeras.search.ModelTrainer.train_model', side_effect=mock_train)
+@patch('autokeras.backend.torch.model_trainer.ModelTrainer.train_model', side_effect=mock_train)
 def test_greedy_searcher_mlp(_, _1, _2):
     train_data, test_data = get_classification_data_loaders_mlp()
     clean_dir(TEST_TEMP_DIR)
@@ -124,7 +124,7 @@ def test_greedy_searcher_mlp(_, _1, _2):
 
 @patch('torch.multiprocessing.get_context', side_effect=MockProcess)
 @patch('autokeras.bayesian.transform', side_effect=simple_transform)
-@patch('autokeras.search.ModelTrainer.train_model', side_effect=mock_train)
+@patch('autokeras.backend.torch.model_trainer.ModelTrainer.train_model', side_effect=mock_train)
 def test_random_searcher(_, _1, _2):
     train_data, test_data = get_classification_data_loaders()
     clean_dir(TEST_TEMP_DIR)
@@ -138,7 +138,7 @@ def test_random_searcher(_, _1, _2):
 
 @patch('torch.multiprocessing.get_context', side_effect=MockProcess)
 @patch('autokeras.bayesian.transform', side_effect=simple_transform)
-@patch('autokeras.search.ModelTrainer.train_model', side_effect=mock_train)
+@patch('autokeras.backend.torch.model_trainer.ModelTrainer.train_model', side_effect=mock_train)
 @patch('autokeras.search.get_system', return_value=Constant.SYS_GOOGLE_COLAB)
 def test_random_searcher_sp(_, _1, _2, _3):
     train_data, test_data = get_classification_data_loaders()
@@ -153,7 +153,7 @@ def test_random_searcher_sp(_, _1, _2, _3):
 
 @patch('torch.multiprocessing.get_context', side_effect=MockProcess)
 @patch('autokeras.bayesian.transform', side_effect=simple_transform)
-@patch('autokeras.search.ModelTrainer.train_model', side_effect=mock_train)
+@patch('autokeras.backend.torch.model_trainer.ModelTrainer.train_model', side_effect=mock_train)
 def test_grid_searcher(_, _1, _2):
     train_data, test_data = get_classification_data_loaders()
     clean_dir(TEST_TEMP_DIR)
@@ -169,7 +169,7 @@ def test_grid_searcher(_, _1, _2):
 
 
 @patch('torch.multiprocessing.get_context', side_effect=MockProcess)
-@patch('autokeras.search.ModelTrainer.train_model', side_effect=mock_train)
+@patch('autokeras.backend.torch.model_trainer.ModelTrainer.train_model', side_effect=mock_train)
 def test_max_acq(_, _2):
     train_data, test_data = get_classification_data_loaders()
     clean_dir(TEST_TEMP_DIR)
@@ -189,7 +189,7 @@ def test_max_acq(_, _2):
 
 
 @patch('torch.multiprocessing.get_context', side_effect=MockProcess)
-@patch('autokeras.search.ModelTrainer.train_model', side_effect=mock_out_of_memory_train)
+@patch('autokeras.backend.torch.model_trainer.ModelTrainer.train_model', side_effect=mock_out_of_memory_train)
 def test_out_of_memory(_, _2):
     train_data, test_data = get_classification_data_loaders()
     clean_dir(TEST_TEMP_DIR)
